Fix based on feedback
