+++ core/src/main/java/com/scalar/db/api/ScanBuilder.java
@@ -25,17 +25,17 @@ import com.scalar.db.api.OperationBuilder.TableBuilder;
 import com.scalar.db.api.OperationBuilder.Where;
 import com.scalar.db.api.OperationBuilder.WhereAnd;
 import com.scalar.db.api.OperationBuilder.WhereOr;
-import com.scalar.db.api.Scan.AndConditionSet;
-import com.scalar.db.api.Scan.OrConditionSet;
 import com.scalar.db.io.Key;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
+import javax.annotation.concurrent.Immutable;
 
 public class ScanBuilder {
 
@@ -473,13 +473,11 @@ public class ScanBuilder {
     @Nullable protected com.scalar.db.api.Consistency consistency;
 
     private BuildableScanAllWithWhere(String namespaceName, String tableName) {
-      this.namespaceName = namespaceName;
-      this.tableName = tableName;
-      this.condition = null;
+      this(namespaceName, tableName, null);
     }
 
     private BuildableScanAllWithWhere(
-        String namespaceName, String tableName, ConditionalExpression condition) {
+        String namespaceName, String tableName, @Nullable ConditionalExpression condition) {
       this.namespaceName = namespaceName;
       this.tableName = tableName;
       this.condition = condition;
@@ -808,8 +806,9 @@ public class ScanBuilder {
 
     private void checkConditionsEmpty() {
       if (!conjunctions.isEmpty()) {
-        throw new UnsupportedOperationException(
-            "This operation is supported only when no conditions are specified at all.");
+        throw new IllegalStateException(
+            "This operation is supported only when no conditions are specified at all. "
+                + "If you want to modify the condition, please use clearConditions() to remove all existing conditions first.");
       }
     }
 
@@ -1081,6 +1080,244 @@ public class ScanBuilder {
       checkNotNull(orConditionSet);
       disjunctions.add(orConditionSet.getConditions());
       return this;
+    }
+  }
+
+  /**
+   * An and-wise set of {@link ConditionalExpression} used for specifying arbitrary conditions in a
+   * {@link Scan} command.
+   */
+  @Immutable
+  public static class AndConditionSet {
+    private final ImmutableSet<ConditionalExpression> conditions;
+
+    private AndConditionSet(ImmutableSet<ConditionalExpression> conditions) {
+      this.conditions = conditions;
+    }
+
+    /**
+     * Returns the set of {@code ConditionalExpression}.
+     *
+     * @return set of {@code ConditionalExpression}
+     */
+    public Set<ConditionalExpression> getConditions() {
+      return conditions;
+    }
+
+    /**
+     * Indicates whether some other object is "equal to" this object. The other object is considered
+     * equal if:
+     *
+     * <ul>
+     *   <li>it is also an {@code AndConditionSet}
+     *   <li>both instances have the same set of {@code ConditionalExpression}
+     * </ul>
+     *
+     * @param o an object to be tested for equality
+     * @return {@code true} if the other object is "equal to" this object otherwise {@code false}
+     */
+    @Override
+    public boolean equals(Object o) {
+      if (o == this) {
+        return true;
+      }
+      if (!(o instanceof AndConditionSet)) {
+        return false;
+      }
+      AndConditionSet other = (AndConditionSet) o;
+      return conditions.equals(other.conditions);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(conditions);
+    }
+
+    @Override
+    public String toString() {
+      return conditions.toString();
+    }
+  }
+
+  /**
+   * An or-wise set of {@link ConditionalExpression} used for specifying arbitrary conditions in a
+   * {@link Scan} command.
+   */
+  @Immutable
+  public static class OrConditionSet {
+    private final ImmutableSet<ConditionalExpression> conditions;
+
+    private OrConditionSet(ImmutableSet<ConditionalExpression> conditions) {
+      this.conditions = conditions;
+    }
+
+    /**
+     * Returns the set of {@code ConditionalExpression}.
+     *
+     * @return set of {@code ConditionalExpression}
+     */
+    public Set<ConditionalExpression> getConditions() {
+      return conditions;
+    }
+
+    /**
+     * Indicates whether some other object is "equal to" this object. The other object is considered
+     * equal if:
+     *
+     * <ul>
+     *   <li>it is also an {@code OrConditionSet}
+     *   <li>both instances have the same set of {@code ConditionalExpression}
+     * </ul>
+     *
+     * @param o an object to be tested for equality
+     * @return {@code true} if the other object is "equal to" this object otherwise {@code false}
+     */
+    @Override
+    public boolean equals(Object o) {
+      if (o == this) {
+        return true;
+      }
+      if (!(o instanceof OrConditionSet)) {
+        return false;
+      }
+      OrConditionSet other = (OrConditionSet) o;
+      return conditions.equals(other.conditions);
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(conditions);
+    }
+
+    @Override
+    public String toString() {
+      return conditions.toString();
+    }
+  }
+
+  public static class ConditionSetBuilder {
+
+    /**
+     * Returns a builder object to build a {@code AndConditionSet} or {@code OrConditionSet}.
+     *
+     * @param condition a conditional expression to build a {@code AndConditionSet} or {@code
+     *     OrConditionSet}.
+     * @return a builder object
+     */
+    public static AndOrConditionSetBuilder condition(ConditionalExpression condition) {
+      return new AndOrConditionSetBuilder(ImmutableSet.of(condition));
+    }
+
+    /**
+     * Returns a builder object to build a {@code AndConditionSet}.
+     *
+     * @param conditions a set of conditional expressions to build a {@code AndConditionSet}.
+     * @return a builder object
+     */
+    public static BuildableAndConditionSet andConditionSet(Set<ConditionalExpression> conditions) {
+      return new BuildableAndConditionSet(conditions);
+    }
+
+    /**
+     * Returns a builder object to build a {@code OrConditionSet}.
+     *
+     * @param conditions a set of conditional expressions to build a {@code OrConditionSet}.
+     * @return a builder object
+     */
+    public static BuildableOrConditionSet orConditionSet(Set<ConditionalExpression> conditions) {
+      return new BuildableOrConditionSet(conditions);
+    }
+  }
+
+  public static class AndOrConditionSetBuilder {
+
+    private final Set<ConditionalExpression> conditions;
+
+    private AndOrConditionSetBuilder(Set<ConditionalExpression> conditions) {
+      this.conditions = new HashSet<>();
+      this.conditions.addAll(conditions);
+    }
+
+    /**
+     * Adds a conditional expression for a {@code AndConditionSet}.
+     *
+     * @param condition a conditional expression for a {@code AndConditionSet}.
+     * @return a builder object
+     */
+    public BuildableAndConditionSet and(ConditionalExpression condition) {
+      conditions.add(condition);
+      return new BuildableAndConditionSet(conditions);
+    }
+
+    /**
+     * Adds a conditional expression for a {@code OrConditionSet}.
+     *
+     * @param condition a conditional expression for a {@code OrConditionSet}.
+     * @return a builder object
+     */
+    public BuildableOrConditionSet or(ConditionalExpression condition) {
+      conditions.add(condition);
+      return new BuildableOrConditionSet(conditions);
+    }
+  }
+
+  public static class BuildableAndConditionSet {
+
+    private final Set<ConditionalExpression> conditions;
+
+    private BuildableAndConditionSet(Set<ConditionalExpression> conditions) {
+      this.conditions = new HashSet<>();
+      this.conditions.addAll(conditions);
+    }
+
+    /**
+     * Adds a conditional expression for a {@code AndConditionSet}.
+     *
+     * @param condition a conditional expression for a {@code AndConditionSet}.
+     * @return a builder object
+     */
+    public BuildableAndConditionSet and(ConditionalExpression condition) {
+      conditions.add(condition);
+      return this;
+    }
+
+    /**
+     * Builds an and-wise condition set with the specified conditional expressions.
+     *
+     * @return an and-wise condition set
+     */
+    public AndConditionSet build() {
+      return new AndConditionSet(ImmutableSet.copyOf(conditions));
+    }
+  }
+
+  public static class BuildableOrConditionSet {
+
+    private final Set<ConditionalExpression> conditions;
+
+    private BuildableOrConditionSet(Set<ConditionalExpression> conditions) {
+      this.conditions = new HashSet<>();
+      this.conditions.addAll(conditions);
+    }
+
+    /**
+     * Adds a conditional expression for a {@code OrConditionSet}.
+     *
+     * @param condition a conditional expression for a {@code OrConditionSet}.
+     * @return a builder object
+     */
+    public BuildableOrConditionSet or(ConditionalExpression condition) {
+      conditions.add(condition);
+      return this;
+    }
+
+    /**
+     * Builds an or-wise condition set with the specified conditional expressions.
+     *
+     * @return an or-wise condition set
+     */
+    public OrConditionSet build() {
+      return new OrConditionSet(ImmutableSet.copyOf(conditions));
     }
   }
